requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
openpyxl
python-docx
langchain-community
beautifulsoup4
requests
playwright